fixed code to be flake8 and pylint compliant
