+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 
-exclude = .git,__pycache__,docs/,requirements/,conf/settings/, .env/, **/migrations/
+exclude = .git,__pycache__,docs/,requirements/,conf/settings/, .env/, **/migrations/, setup.py, build/
 show-source = True
 statistics = True
 count = True
