--- src/kong/kong_client.py
@@ -1,3 +1,0 @@
-from kong.KongAdminClient import KongAdminClient
-
-client = KongAdminClient
