+++ tests/api_admin_tests.py
@@ -4,7 +4,7 @@ from faker import Faker
 
 from kong.providers import ApiDataProvider
 
-from kong.ApiAdminClient import ApiAdminClient
+from kong.kong_clients import ApiAdminClient
 from kong.structures import ApiData
 
 
@@ -61,7 +61,8 @@ class ApiAdminClientTest(unittest.TestCase):
         self.kong_admin_url = self.faker.url()
         self.apis_endpoint = self.kong_admin_url + 'apis/'
 
-        self.api_admin_client = ApiAdminClient(self.kong_admin_url, session=self.requests_mock.session())
+        self.api_admin_client = ApiAdminClient(self.kong_admin_url,
+                                               _session=self.requests_mock.session())
 
     def test_api_admin_create(self):
         """
@@ -70,7 +71,9 @@ class ApiAdminClientTest(unittest.TestCase):
         """
 
         # Exercise
-        api_data = self.api_admin_client.create(self.api_name, self.api_upstream_url, uris=self.api_uris)
+        api_data = self.api_admin_client.create(self.api_name,
+                                                self.api_upstream_url,
+                                                uris=self.api_uris)
 
         # Verify
         self.assertEqual(api_data['name'], self.api_name)
@@ -98,7 +101,9 @@ class ApiAdminClientTest(unittest.TestCase):
             as normal create
         """
         # Setup
-        orig_data = ApiData(name=self.api_name, upstream_url=self.api_upstream_url, uris=self.api_uris)
+        orig_data = ApiData(name=self.api_name,
+                            upstream_url=self.api_upstream_url,
+                            uris=self.api_uris)
 
         # Exercise
         self.api_admin_client.create(orig_data)
@@ -117,7 +122,6 @@ class ApiAdminClientTest(unittest.TestCase):
         self.api_admin_client.delete(self.api_name)
 
         # Verify
-        expected_data = {}
         api_endpoint = self.apis_endpoint + self.api_name
         self.session_mock.delete.assert_called_once_with(api_endpoint)
 
@@ -132,13 +136,14 @@ class ApiAdminClientTest(unittest.TestCase):
         self.api_admin_client.delete(self.api_kong_id)
 
         # Verify
-        expected_data = {}
         api_endpoint = self.apis_endpoint + self.api_kong_id
         self.session_mock.delete.assert_called_once_with(api_endpoint)
 
     def test_api_admin_update(self):
         # Setup
-        api_data = self.api_admin_client.create(self.api_name, self.api_upstream_url, uris=self.api_uris)
+        api_data = self.api_admin_client.create(self.api_name,
+                                                self.api_upstream_url,
+                                                uris=self.api_uris)
         new_uri = self.faker.api_path()
 
         # Exercise
@@ -278,7 +283,8 @@ class ApiAdminClientTest(unittest.TestCase):
 
         # Verify
         self.assertRaisesRegex(KeyError, r"unknown field",
-                               lambda: self.api_admin_client.update(self.api_data['name'], **self.api_data))
+                               lambda: self.api_admin_client
+                               .update(self.api_data['name'], **self.api_data))
 
     def test_update_not_existing_api(self):
         # Setup
@@ -287,7 +293,8 @@ class ApiAdminClientTest(unittest.TestCase):
 
         # Verify
         self.assertRaisesRegex(NameError, r"not found",
-                               lambda: self.api_admin_client.update(self.api_data['name'], **self.api_data))
+                               lambda: self.api_admin_client
+                               .update(self.api_data['name'], **self.api_data))
 
     def test_update_internal_server_error(self):
         # Setup
@@ -296,7 +303,8 @@ class ApiAdminClientTest(unittest.TestCase):
 
         # Verify
         self.assertRaisesRegex(Exception, r'internal server error',
-                               lambda: self.api_admin_client.update(self.api_data['name'], **self.api_data))
+                               lambda: self.api_admin_client
+                               .update(self.api_data['name'], **self.api_data))
 
     def test_list_internal_server_error(self):
         # Setup
